# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export type GoalType = typeof GoalType[keyof typeof GoalType];


export const GoalType = {
  daily: 'daily',
  scheduled: 'scheduled',
} as const;

export type GoalPriority = typeof GoalPriority[keyof typeof GoalPriority];


export const GoalPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  pending: 'pending',
  done: 'done',
} as const;

export interface Goal {
  id: string;
  userId: string;
  title: string;
  /** @nullable */
  intent?: string | null;
  type: GoalType;
  priority: GoalPriority;
  startTime: string;
  durationMinutes: number;
  /** @nullable */
  scheduledDate?: string | null;
  status: GoalStatus;
  createdAt: string;
  updatedAt: string;
}

export type GoalInputType = typeof GoalInputType[keyof typeof GoalInputType];


export const GoalInputType = {
  daily: 'daily',
  scheduled: 'scheduled',
} as const;

export type GoalInputPriority = typeof GoalInputPriority[keyof typeof GoalInputPriority];


export const GoalInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  intent?: string;
  type: GoalInputType;
  priority: GoalInputPriority;
  startTime: string;
  /** @minimum 1 */
  durationMinutes: number;
  scheduledDate?: string;
}

export type GoalUpdateType = typeof GoalUpdateType[keyof typeof GoalUpdateType];


export const GoalUpdateType = {
  daily: 'daily',
  scheduled: 'scheduled',
} as const;

export type GoalUpdatePriority = typeof GoalUpdatePriority[keyof typeof GoalUpdatePriority];


export const GoalUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type GoalUpdateStatus = typeof GoalUpdateStatus[keyof typeof GoalUpdateStatus];


export const GoalUpdateStatus = {
  pending: 'pending',
  done: 'done',
} as const;

export interface GoalUpdate {
  /** @minLength 1 */
  title?: string;
  intent?: string;
  type?: GoalUpdateType;
  priority?: GoalUpdatePriority;
  startTime?: string;
  /** @minimum 1 */
  durationMinutes?: number;
  scheduledDate?: string;
  status?: GoalUpdateStatus;
}

export interface AppUsageEntry {
  id: string;
  userId: string;
  appName: string;
  /** @nullable */
  category?: string | null;
  durationMinutes: number;
  loggedAt: string;
  createdAt: string;
}

export interface AppUsageEntryInput {
  /** @minLength 1 */
  appName: string;
  category?: string;
  /** @minimum 1 */
  durationMinutes: number;
  loggedAt?: string;
}

export interface AppUsageSummaryItem {
  appName: string;
  /** @nullable */
  category?: string | null;
  totalMinutes: number;
}

export type FocusSessionStatus = typeof FocusSessionStatus[keyof typeof FocusSessionStatus];


export const FocusSessionStatus = {
  running: 'running',
  paused: 'paused',
  completed: 'completed',
  stopped: 'stopped',
} as const;

export interface FocusSession {
  id: string;
  userId: string;
  /** @nullable */
  intention?: string | null;
  plannedDurationMinutes: number;
  blockedApps: string[];
  status: FocusSessionStatus;
  startedAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface FocusSessionInput {
  intention?: string;
  /** @minimum 1 */
  plannedDurationMinutes: number;
  blockedApps: string[];
}

export type FocusSessionEndInputStatus = typeof FocusSessionEndInputStatus[keyof typeof FocusSessionEndInputStatus];


export const FocusSessionEndInputStatus = {
  completed: 'completed',
  stopped: 'stopped',
} as const;

export interface FocusSessionEndInput {
  status: FocusSessionEndInputStatus;
}

export interface BlockedApp {
  id: string;
  userId: string;
  appName: string;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface BlockedAppInput {
  /** @minLength 1 */
  appName: string;
  category?: string;
}

export type UserSettingsAppearance = typeof UserSettingsAppearance[keyof typeof UserSettingsAppearance];


export const UserSettingsAppearance = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserSettings {
  userId: string;
  notificationsEnabled: boolean;
  dailyReminderEnabled: boolean;
  focusReminderEnabled: boolean;
  appearance: UserSettingsAppearance;
  analyticsOptIn: boolean;
  usageTrackingOptIn: boolean;
  updatedAt: string;
}

export type UserSettingsUpdateAppearance = typeof UserSettingsUpdateAppearance[keyof typeof UserSettingsUpdateAppearance];


export const UserSettingsUpdateAppearance = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UserSettingsUpdate {
  notificationsEnabled?: boolean;
  dailyReminderEnabled?: boolean;
  focusReminderEnabled?: boolean;
  appearance?: UserSettingsUpdateAppearance;
  analyticsOptIn?: boolean;
  usageTrackingOptIn?: boolean;
}