# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals for the current user
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        title:
          type: string
        intent:
          type: ["string", "null"]
        type:
          type: string
          enum: [daily, scheduled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startTime:
          type: string
        durationMinutes:
          type: integer
        scheduledDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, done]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - type
        - priority
        - startTime
        - durationMinutes
        - status
        - createdAt
        - updatedAt
    GoalInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        intent:
          type: string
        type:
          type: string
          enum: [daily, scheduled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startTime:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        scheduledDate:
          type: string
      required: [title, type, priority, startTime, durationMinutes]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        intent:
          type: string
        type:
          type: string
          enum: [daily, scheduled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        startTime:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        scheduledDate:
          type: string
        status:
          type: string
          enum: [pending, done]
...