# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalPriority } from './goalPriority';
import type { GoalStatus } from './goalStatus';
import type { GoalType } from './goalType';

export interface Goal {
  id: string;
  userId: string;
  title: string;
  /** @nullable */
  intent?: string | null;
  type: GoalType;
  priority: GoalPriority;
  startTime: string;
  durationMinutes: number;
  /** @nullable */
  scheduledDate?: string | null;
  status: GoalStatus;
  createdAt: Date;
  updatedAt: Date;
}